+++ example/lib/feedback_manager/feedback_manager_screen.dart
@@ -9,9 +9,9 @@ class FeedbackManagerScreen extends CubeWidget<FeedbackManagerCube> {
       appBar: AppBar(
         title: Text('FeedBackManager'),
       ),
-      body: FeedBackManager(
+      body: CFeedBackManager(
         dialogControllers: [
-          DialogController<String>(
+          CDialogController<String>(
             observable: cube.dialogControl,
             dismissible: false,
             builder: (data, context) => Center(
@@ -25,7 +25,7 @@ class FeedbackManagerScreen extends CubeWidget<FeedbackManagerCube> {
           ),
         ],
         bottomSheetControllers: [
-          BottomSheetController<String>(
+          CBottomSheetController<String>(
             dismissible: false,
             observable: cube.bottomSheetControl,
             builder: (data, context) => Container(
@@ -37,7 +37,7 @@ class FeedbackManagerScreen extends CubeWidget<FeedbackManagerCube> {
           ),
         ],
         snackBarControllers: [
-          SnackBarController(
+          CSnackBarController(
             observable: cube.snackBarControl,
             builder: (data, context) => SnackBar(content: Text(data)),
           ),